+++ relay-server/src/actors/project_cache.rs
@@ -8,7 +8,7 @@ use failure::Fail;
 use futures::{future, Future};
 use serde::{Deserialize, Serialize};
 
-use relay_common::{metric, ProjectId};
+use relay_common::{metric, ProjectKey};
 use relay_config::{Config, RelayMode};
 use relay_redis::RedisPool;
 
@@ -40,7 +40,7 @@ struct ProjectEntry {
 
 pub struct ProjectCache {
     config: Arc<Config>,
-    projects: HashMap<ProjectId, ProjectEntry>,
+    projects: HashMap<ProjectKey, ProjectEntry>,
 
     local_source: Addr<LocalProjectSource>,
     upstream_source: Addr<UpstreamProjectSource>,
@@ -113,9 +113,19 @@ impl Actor for ProjectCache {
     }
 }
 
+/// Resolves the project with the given identifier.
+///
+/// The returned `Project` is an actor that synchronizes state access internally. When it is fetched
+/// for the first time, its state is unpopulated. Only when `GetProjectState` is sent to the project
+/// for the first time, it starts to resolve the state from one of the sources.
+///
+/// If the optional `public_key` is set, then the public keys of the project are checked for a
+/// redirect. If a redirect is detected, then the target project is resolved and returned instead.
+///
+/// **Note** that due to redirects, the returned project may have a different identifier.
 #[derive(Debug, Deserialize, Serialize)]
 pub struct GetProject {
-    pub id: ProjectId,
+    pub public_key: ProjectKey,
 }
 
 impl Message for GetProject {
@@ -126,16 +136,18 @@ impl Handler<GetProject> for ProjectCache {
     type Result = Addr<Project>;
 
     fn handle(&mut self, message: GetProject, context: &mut Context<Self>) -> Self::Result {
+        let GetProject { public_key } = message;
         let config = self.config.clone();
         metric!(histogram(RelayHistograms::ProjectStateCacheSize) = self.projects.len() as u64);
-        match self.projects.entry(message.id) {
+
+        match self.projects.entry(public_key) {
             Entry::Occupied(entry) => {
                 metric!(counter(RelayCounters::ProjectCacheHit) += 1);
                 entry.get().project.clone()
             }
             Entry::Vacant(entry) => {
                 metric!(counter(RelayCounters::ProjectCacheMiss) += 1);
-                let project = Project::new(message.id, config, context.address()).start();
+                let project = Project::new(public_key, config, context.address()).start();
                 entry.insert(ProjectEntry {
                     last_updated_at: Instant::now(),
                     project: project.clone(),
@@ -146,9 +158,19 @@ impl Handler<GetProject> for ProjectCache {
     }
 }
 
-#[derive(Clone, Copy)]
+/// Fetches a project state from one of the available sources.
+///
+/// The project state is resolved in the following precedence:
+///
+///  1. Local file system
+///  2. Redis cache (processing mode only)
+///  3. Upstream (managed and processing mode only)
+///
+/// Requests to the upstream are performed via `UpstreamProjectSource`, which internally batches
+/// individual requests.
+#[derive(Clone)]
 pub struct FetchProjectState {
-    pub id: ProjectId,
+    pub public_key: ProjectKey,
 }
 
 pub struct ProjectStateResponse {
@@ -176,8 +198,9 @@ impl Message for FetchProjectState {
     type Result = Result<ProjectStateResponse, ()>;
 }
 
+#[derive(Clone, Debug)]
 pub struct FetchOptionalProjectState {
-    pub id: ProjectId,
+    pub public_key: ProjectKey,
 }
 
 impl Message for FetchOptionalProjectState {
@@ -188,7 +211,8 @@ impl Handler<FetchProjectState> for ProjectCache {
     type Result = Response<ProjectStateResponse, ()>;
 
     fn handle(&mut self, message: FetchProjectState, _context: &mut Self::Context) -> Self::Result {
-        if let Some(mut entry) = self.projects.get_mut(&message.id) {
+        let FetchProjectState { public_key } = message;
+        if let Some(mut entry) = self.projects.get_mut(&public_key) {
             // Bump the update time of the project in our hashmap to evade eviction. Eviction is a
             // sequential scan over self.projects, so this needs to be as fast as possible and
             // probably should not involve sending a message to each Addr<Project> (this is the
@@ -209,7 +233,7 @@ impl Handler<FetchProjectState> for ProjectCache {
 
         let fetch_local = self
             .local_source
-            .send(FetchOptionalProjectState { id: message.id })
+            .send(FetchOptionalProjectState { public_key })
             .map_err(|_| ());
 
         let future = fetch_local.and_then(move |response| {
@@ -246,7 +270,7 @@ impl Handler<FetchProjectState> for ProjectCache {
             let fetch_redis: ResponseFuture<_, _> = if let Some(ref redis_source) = redis_source {
                 Box::new(
                     redis_source
-                        .send(FetchOptionalProjectState { id: message.id })
+                        .send(FetchOptionalProjectState { public_key })
                         .map_err(|_| ()),
                 )
             } else {