+++ relay-server/src/actors/project_redis.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use actix::prelude::*;
 use failure::Fail;
-use relay_common::{LogError, ProjectId};
+use relay_common::{LogError, ProjectKey};
 use relay_config::Config;
 use relay_redis::{RedisError, RedisPool};
 
@@ -40,11 +40,11 @@ impl RedisProjectSource {
         RedisProjectSource { config, redis }
     }
 
-    fn get_config(&self, id: ProjectId) -> Result<Option<ProjectState>, RedisProjectError> {
+    fn get_config(&self, key: ProjectKey) -> Result<Option<ProjectState>, RedisProjectError> {
         let mut command = relay_redis::redis::cmd("GET");
 
         let prefix = self.config.projectconfig_cache_prefix();
-        command.arg(format!("{}:{}", prefix, id));
+        command.arg(format!("{}:{}", prefix, key));
 
         let raw_response_opt: Option<String> = command
             .query(&mut self.redis.client()?.connection())
@@ -79,7 +79,7 @@ impl Handler<FetchOptionalProjectState> for RedisProjectSource {
         message: FetchOptionalProjectState,
         _ctx: &mut Self::Context,
     ) -> Self::Result {
-        match self.get_config(message.id) {
+        match self.get_config(message.public_key) {
             Ok(x) => x.map(ProjectState::sanitize).map(Arc::new),
             Err(e) => {
                 log::error!("Failed to fetch project from Redis: {}", LogError(&e));