+++ relay-server/src/actors/project_upstream.rs
@@ -10,7 +10,7 @@ use futures::{future, future::Shared, sync::oneshot, Future};
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 
-use relay_common::{metric, LogError, ProjectId, RetryBackoff};
+use relay_common::{metric, LogError, ProjectKey, RetryBackoff};
 use relay_config::Config;
 
 use crate::actors::project::ProjectState;
@@ -19,17 +19,33 @@ use crate::actors::upstream::{RequestPriority, SendQuery, UpstreamQuery, Upstrea
 use crate::metrics::{RelayCounters, RelayHistograms, RelayTimers};
 use crate::utils::{self, ErrorBoundary};
 
+#[macro_use]
+mod _macro {
+    /// The current version of the project states endpoint.
+    ///
+    /// Only this version is supported by Relay. All other versions are forwarded to the Upstream.
+    /// The endpoint version is added as `version` query parameter to every outgoing request.
+    #[macro_export]
+    macro_rules! project_states_version {
+        () => {
+            2
+        };
+    }
+}
+
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GetProjectStates {
-    pub projects: Vec<ProjectId>,
+    pub public_keys: Vec<ProjectKey>,
     #[serde(default)]
     pub full_config: bool,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
 pub struct GetProjectStatesResponse {
-    pub configs: HashMap<ProjectId, ErrorBoundary<Option<ProjectState>>>,
+    #[serde(default)]
+    pub configs: HashMap<ProjectKey, ErrorBoundary<Option<ProjectState>>>,
 }
 
 impl UpstreamQuery for GetProjectStates {
@@ -40,7 +56,10 @@ impl UpstreamQuery for GetProjectStates {
     }
 
     fn path(&self) -> Cow<'static, str> {
-        Cow::Borrowed("/api/0/relays/projectconfigs/")
+        Cow::Borrowed(concat!(
+            "/api/0/relays/projectconfigs/?version=",
+            project_states_version!()
+        ))
     }
 
     fn priority() -> RequestPriority {
@@ -87,7 +106,7 @@ pub struct UpstreamProjectSource {
     backoff: RetryBackoff,
     config: Arc<Config>,
     upstream: Addr<UpstreamRelay>,
-    state_channels: HashMap<ProjectId, ProjectStateChannel>,
+    state_channels: HashMap<ProjectKey, ProjectStateChannel>,
 }
 
 impl UpstreamProjectSource {
@@ -175,8 +194,8 @@ impl UpstreamProjectSource {
                         channels_batch.len() as u64
                 );
 
-                let request = GetProjectStates {
-                    projects: channels_batch.keys().copied().collect(),
+                let query = GetProjectStates {
+                    public_keys: channels_batch.keys().copied().collect(),
                     full_config: self.config.processing_enabled(),
                 };
 
@@ -184,7 +203,7 @@ impl UpstreamProjectSource {
                 metric!(counter(RelayCounters::ProjectStateRequest) += 1);
 
                 self.upstream
-                    .send(SendQuery(request))
+                    .send(SendQuery(query))
                     .map_err(ProjectError::ScheduleFailed)
                     .map(move |response| (channels_batch, response))
             })
@@ -213,14 +232,14 @@ impl UpstreamProjectSource {
                                 histogram(RelayHistograms::ProjectStateReceived) =
                                     response.configs.len() as u64
                             );
-                            for (id, channel) in channels_batch {
+                            for (key, channel) in channels_batch {
                                 let state = response
                                     .configs
-                                    .remove(&id)
+                                    .remove(&key)
                                     .unwrap_or(ErrorBoundary::Ok(None))
                                     .unwrap_or_else(|error| {
                                         let e = LogError(error);
-                                        log::error!("error fetching project state {}: {}", id, e);
+                                        log::error!("error fetching project state {}: {}", key, e);
                                         Some(ProjectState::err())
                                     })
                                     .unwrap_or_else(ProjectState::missing);
@@ -309,7 +328,7 @@ impl Handler<FetchProjectState> for UpstreamProjectSource {
         // channel for our current `message.id`.
         let channel = self
             .state_channels
-            .entry(message.id)
+            .entry(message.public_key)
             .or_insert_with(|| ProjectStateChannel::new(query_timeout));
 
         Box::new(